env: typescript non strict
